Seems like it should work, but recipes aren't working as you'd think they would...

+++ src/LifeSupportStatusMonitor.cs
@@ -112,7 +112,7 @@ namespace Nerm.Colonization
                                 new DialogGUILabel(resourceTransfer.TargetVessel?.GetDisplayName())),
                             new DialogGUIHorizontalLayout(TextAnchor.MiddleLeft,
                                 new DialogGUIButton("Start", resourceTransfer.StartTransfer, () => resourceTransfer.TargetVessel != null && !resourceTransfer.IsTransferUnderway, dismissOnSelect: false),
-                                new DialogGUISlider(() => resourceTransfer.TransferPercent, 0, 1, false, 100, 20, null))));
+                                new DialogGUISlider(() => (float)resourceTransfer.TransferPercent, 0, 1, false, 100, 20, null))));
                 }
 
                 this.dialog = PopupDialog.SpawnPopupDialog(
